(PC-17816) refactor(profile): move EditButton from pages to components
