+++ src/features/profile/pages/PersonalData/PersonalData.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import styled from 'styled-components/native'
 
 import { useAuthContext } from 'features/auth/AuthContext'
+import { EditButton } from 'features/profile/components/Buttons/EditButton/EditButton'
 import { PageProfileSection } from 'features/profile/components/PageProfileSection/PageProfileSection'
-import { EditButton } from 'features/profile/pages/PersonalData/EditButton'
 import { analytics } from 'libs/firebase/analytics'
 import { SectionRow } from 'ui/components/SectionRow'
 import { Separator } from 'ui/components/Separator'
